move providers to SwitchUser::Provider::ProviderName
fixes error raise by SwitchUserController#provider_class
autoload wasn't finding the constant Provider
Provider wasn't namespaced under SwitchUser, so it was poluting the namespace
