+++ lib/switch_user/provider/authlogic.rb
@@ -1,0 +1,24 @@
+module SwitchUser
+  module Provider
+    class Authlogic
+      def initialize(controller)
+        @controller = controller
+      end
+
+      def login(user, scope = nil)
+        UserSession.create(user)
+      end
+
+      def logout(scope = nil)
+        @controller.current_user_session.destroy
+      end
+
+      def current_user(scope = nil)
+        result = UserSession.find
+        if result
+          result.record
+        end
+      end
+    end
+  end
+end
