+++ lib/switch_user/provider/restful_authentication.rb
@@ -1,0 +1,21 @@
+module SwitchUser
+  module Provider
+    class RestfulAuthentication
+      def initialize(controller)
+        @controller = controller
+      end
+
+      def login(user, scope = nil)
+        @controller.current_user = user
+      end
+
+      def logout(scope = nil)
+        @controller.logout_killing_session!
+      end
+
+      def current_user(scope = nil)
+        @controller.current_user
+      end
+    end
+  end
+end
